+++ app/__main__.py
@@ -1,20 +1,51 @@
 """
 Micro Service with Ports and Adapters Architecture
 ==================================================
-The application's entrance point. Loads and configures the adapters to be used
-by the service.
+The application's entry point. Loads and configures the adapters to be used
+by the application.
 
 __license__ =  MIT
 __author__ = 'Lucas Góes'
 __email__ = 'lucas.rd.goes@gmail.com'
 """
 
+import time
+import logging
+
+from . import ports, adapters, settings
 from .version import __version__
 
 
 def main():
 	"""The application's main function."""
-	pass
+	app_config = settings.ApplicationConfig()
+
+	# Configures application logger.
+	logging.basicConfig(
+		level=app_config.logger_level,
+		format='%(asctime) -19s | %(levelname) -8s | %(threadName) -10s |'
+			   ' %(funcName) -16s | %(message)s'
+	)
+
+	logger = logging.getLogger(app_config.logger_name)
+	logger.info('Started sample book managing application v{0}.' \
+				.format(__version__))
+
+	# Creates CommandBus.
+	command_bus = ports.CommandBus()
+
+	# Configures and starts adapters.
+	driver_adapter = app_config.driver_adapter
+	if driver_adapter == 'mqtt':
+		driver_config = settings.MqttDriverConfig()
+		driver_adapter = adapters.MqttAdapter(driver_config, command_bus)
+
+	# Runs driver adapter and awaits for any external interruption.
+	try:
+		driver_adapter.start()
+	except KeyboardInterrupt:
+		driver_adapter.stop()
+		logger.info('The application has been interrupted.')
 
 
 if __name__ == '__main__':