+++ app/adapters/__init__.py
@@ -1,10 +1,13 @@
 """
 Adapters
 ========
-
-- Actors interact with the applications ports through adapters using specific
-technologies;
-- An adapter 
+	The adapters act as middlemen between technology and the application's
+ports. There are driver adapters, and driven adapters. Driver adapters
+interact with the application through its commands while the driven adapters
+implement the application's driven ports to be used by the application.
 """
 
-__all__ = []
+from .mqtt import MqttAdapter
+
+
+__all__ = ['mqtt']