+++ app/ports.py
@@ -1,67 +1,124 @@
-""""""
+"""
+Ports
+=====
+	It presents commands for usage of driver adapters to communicate with the
+application, an API (application programming interface) of sorts to interface
+with it in a technology agnostic way.
+	It also has interfaces for driven adapters to implement so that it can
+communicate with them, like Java's SPI (service provider interface).
+"""
 
-class BookCommand(object):
-	"""Generic command that handles books. Further command classes should
-	extend this class."""
-	def __init__(self):
-		"""BookCommand's constructor."""
-		pass
+from collections import namedtuple
 
-
-class RegisterBookCommand(BookCommand):
-	"""Command for registering a new book at the library."""
-	def __init__(self, author: str, isbn: str, name: str):
-		"""RegisterBookCommand's constructor."""
-		self.author = author
-		self.isbn = isbn
-		self.name = name
+from .errors import CommandNotFoundError
 
 
-class BorrowBookCommand(BookCommand):
-	"""Command for borrowing a book from the library."""
-	def __init__(self, isbn: str):
-		"""BorrowBookCommand's constructor."""
-		self.isbn = isbn
+"""
+	These are the application's commands or API. Simple (key: value) objects
+that holds the necessary information for the commands to be executed.
+	The commands are the application's ports for driver adapters, so they
+should always be named by actions that represent interaction with the service.
+	For each command there is a unique handler that can resolve it.
+"""
+RegisterBookCommand = namedtuple(
+	'RegisterBookCommand', ['author', 'isbn', 'name'])
+
+BorrowBookCommand = namedtuple('BorrowBookCommand', ['isbn'])
+
+ReturnBookCommand = namedtuple('ReturnBookCommand', ['isbn'])
 
 
-class ReturnBookCommand(BookCommand):
-	"""Command for returning a book to the library."""
-	def __init__(self, isbn: str):
-		"""ReturnBookCommand's constructor."""
-		self.isbn = isbn
+def register_book_command_handler(command: RegisterBookCommand):
+	"""Handler for the registering of a new book at the library.
+
+	Params
+	------
+	command: RegisterBookCommand -- the expected command to register a book
+	"""
+	print('Registering book: {0}'.format(command))
 
 
-def register_book_handler(command: RegisterBookCommand):
-	"""Handler for the registering of a new book at the library."""
-	print('Hello 1')
-	return True
+def borrow_book_command_handler(command: BorrowBookCommand):
+	"""Handler for the borrowing of a book from the library.
+
+	Params
+	------
+	command: BorrowBookCommand -- the expected command to borrow a book
+	"""
+	print('Borrowing book: {0}'.format(command))
 
 
-def borrow_book_handler(command: BorrowBookCommand):
-	"""Handler for the borrowing of a book from the library."""
-	print('Hello 2')
-	return True
+def return_book_command_handler(command: ReturnBookCommand):
+	"""Handler for the returning of a book to the library.
+
+	Params
+	------
+	command: ReturnBookCommand -- the expected command to return a book
+	"""
+	print('Returning book: {0}'.format(command))
 
 
-def return_book_handler(command: ReturnBookCommand):
-	"""Handler for the returning of a book to the library."""
-	print('Hello 3')
-	return True
-
-
-class StandardCommandBus(object):
-	"""The command bus is responsible for all of the commands to be executed.
+class CommandBus(object):
+	"""The command bus was developed following the Command Bus design pattern.
+	It is responsible for the execution of commands by running the
+	corresponding handler of a chosen command.
 	"""
 	def __init__(self):
-		"""StandardCommandBus's constructor. Creates a command handler map to
-		know which handler can work with which command. The relationship is
-		always 1:1."""
+		"""CommandBus's constructor."""
 		self.command_handler_map = {
-			'RegisterBookCommand': register_book_handler,
-			'BorrowBookCommand': borrow_book_handler,
-			'ReturnBookCommand': return_book_handler
+			'RegisterBookCommand': register_book_command_handler,
+			'BorrowBookCommand': borrow_book_command_handler,
+			'ReturnBookCommand': return_book_command_handler
 		}
 
-	def dispatch(self, command: BookCommand):
-		handler = self.command_handler_map[command.__class__.__name__]
-		return handler(command)
+	def dispatch(self, command):
+		"""CommandBus's dispatcher, responsible for mapping the chosen command
+		to a handler and executing it.
+
+		Params
+		------
+		command -- an existing command of the application's API 
+		"""
+		command_name = command.__class__.__name__
+		handler = self.command_handler_map.get(command_name)
+
+		if handler is not None:
+			handler(command)
+		else:
+			raise CommandNotFoundError('the command \'{0}\' doesn\'t exist.' \
+									   .format(command_name))
+
+
+"""
+	These are the interfaces for the driven adapters to implement, or the
+application's SPI.
+"""
+class BookStorage(object):
+	"""BookStorage is an interface for driven adapters that communicate with
+	databases."""
+	def register(self, book):
+		"""Method to be implemented to register books to the application.
+
+		Params
+		------
+		book -- the book to be inserted on the application
+		"""
+		pass
+
+	def borrow(self, isbn):
+		"""Method to be implemented to borrow books from the application.
+
+		Params
+		------
+		isbn -- the isbn from the book to be borrowed from the application
+		"""
+		pass
+
+	def return_(self, isbn):
+		"""Method to be implemented to return books to the application.
+
+		Params
+		------
+		isbn -- the isbn from the book to be returned to the application
+		"""
+		pass